requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
httpx>=0.25.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
